+++ src/blocks/section-header/index.js
@@ -1,0 +1,13 @@
+/**
+ * Section Header Block
+ */
+
+import { registerBlockType } from '@wordpress/blocks';
+import metadata from './block.json';
+import Edit from './edit';
+import './style.scss';
+
+registerBlockType( metadata.name, {
+	edit: Edit,
+	save: () => null,
+} );